+++ Lab3/Java/Lab3/src/task1/Task1.java
@@ -47,7 +47,7 @@ public class Task1 {
         }
     }
 
-    public static int next(int last, int a, int c, long m) {
+    public static int next(long last, int a, int c, long m) {
         return (int) ((a * last + c) % m);
     }
